bugfix: Internal Server Error upon owner login
When a owner tried to login, internal server error occured, because the Dao method called by OwnerHomeServlet executed an SQL query which does not return a ResultSet containing the field "owner_id", but the private method mapDriverFromResultSet() was trying to retrieve the value of that field. It was fixed by adding the owner_id field to all queries in the DAO class for selection.